Fix some problems in resources

+++ app/libraries/helpers.php
@@ -75,10 +75,14 @@ function updateFile($tempLocation, $newFileName, $oldFileName, $type, $folder1=n
 {
     $flag = saveFile($tempLocation, $newFileName, $type, $folder1, $folder2);
     $fileDest = findFileDest($type, $folder1, $folder2, $oldFileName);
-    if(file_exists($fileDest)){
-        return ($flag and unlink($fileDest));
+    if(!empty($oldFileName)){
+        if(file_exists($fileDest)){
+            return ($flag and unlink($fileDest));
+        }else{
+            return false;
+        }
     }else{
-        return false;
+        return true;
     }
 }
 
@@ -173,6 +177,10 @@ function TdeleteFile($fileName, $type, $folder1=null):bool
     }
 }
 
+function checkFileDest($mainDest, $gid, $sid, $fileName){
+    $fileDest = "private_resources/$mainDest/$gid/$sid/$fileName";
+    return file_exists($fileDest);
+}
 
 function getMonthName($monthNumber) {
     return date("F", mktime(0, 0, 0, $monthNumber, 1));
